Readme.md
The Music Player is a final project for Creative Coding course that combines a functional music player with a captivating sound visualizer in the middle. With built-in buttons, users can easily play/pause, switch between songs, and adjust the volume. In addition, the player also offers motion control with Arduino ultrasonic sensors for greater accessibility.

Two different visual scenes are available for users to enjoy and enhance their music experience. The project was completed in just 2 weeks, providing a fast and efficient solution to a common problem: the inability to control your music player when your hands are occupied. With The Music Player's accessibility mode, users can now enjoy their music anytime, anywhere, and in any condition.